Switch to NewGCMWithRandomNonce for fips
